requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dotenv==0.10.3
requests==2.22.0
SQLAlchemy==1.3.11
urllib3==1.25.7
Werkzeug==0.16.0